api/requirements.txt
Flask==3.0.3
Flask_JWT_Extended==4.6.0
flask_sqlalchemy==3.1.1
mysql_connector_repackaged==0.3.1
Werkzeug==3.1.3
